Delete the dead runtime-RAG-refresh lane and the PostHog HogQL pull
Remove refresh.py, runtime_stitching.py, refresh_dataset.py, and
posthog_pull.py plus their tests. The refresh lane had zero production
callers outside itself: the lazy _REFRESH_EXPORT_MODULES table,
__getattr__, and __dir__ in retrieval/__init__.py existed only to serve
it, extract_fit_transitions was called only by refresh.py, and no
loadable artifact can carry the runtime_trace_snapshot RAGSourceRef arm
(build.py constructs only trace_dataset and store.py rejects anything
else), so the Literal narrows to trace_dataset and the tautological kind
guard in store.py drops. posthog_pull was reachable only through
re-exports never used by the CLI or either happy path; the local-file
posthog source (load_posthog_file, normalize_posthog_payload) is
untouched and its source-table wiring keeps passing.

Relocate the journal fixture helpers (_model, _request, _operation,
_decision, _accept, _reserve_candidate, _complete) from refresh_test.py
into runtime_source_test.py, their only remaining consumer, with
_economics().operations[-1] replaced by the identical
_operation(2, SELECTED_CANDIDATE) since routed_completion_economics
passes operations through unchanged. Drop the startup probe for the
removed lazy exports, the release_test doc asserts and the ingest.md
section for the pull API, the pull tests and fake HTTP seam in
posthog_canonical_test.py, and stale error-string prose that referenced
refreshing runtime retrieval evidence. The posthog telemetry dependency
and REQUIRED_CORE_REQUIREMENTS entry stay: that is the product
telemetry client, not this module.
