+++ wmo/simulation/ingest/posthog.py
@@ -5,12 +5,9 @@ from wmo.simulation.ingest.posthog_canonical import (
     load_posthog_file,
     normalize_posthog_payload,
 )
-from wmo.simulation.ingest.posthog_pull import PostHogPullRequest, pull_posthog_traces
 
 __all__ = [
     "PostHogPullError",
-    "PostHogPullRequest",
     "load_posthog_file",
     "normalize_posthog_payload",
-    "pull_posthog_traces",
 ]
